+++ GdevFallingTiles/BoxRenderer.h
@@ -7,14 +7,14 @@ public:
 	int pixelsPerFill;
 	float fillsPerSecond;
 
-	virtual void OnUpdate(float _deltaTime) override;
+	virtual void OnUpdate(double _deltaTime) override;
 	virtual void OnDraw(sf::RenderWindow& _window, LeaVec2 _screenPosition) override;
 
 
 private:
 	int sizeX, sizeY;
-	float lastFillTime;
-	float timer;
+	double lastFillTime;
+	double timer;
 	bool canFill;
 	LeaVec2 GetRandomPoint();
 	LeaVec2 GetPoint(double _uLerp, double _vLerp);
